Update README.md
This project will predict the likelihood of a customer going default on a loan or not. It will apply a RandomForestClassifier and will maximize the accuracy of the model, but also try to minimize the recall in order to make it sustainable for banks due to not having a lot of false negatives.